requirements.txt
# TIGOTÀ Sistema di Timbratura - Requirements Produzione
# Versione: 3.1.0 Elite

# === CORE GRAPHICS ===
Pillow>=10.0.0         # Image processing

# === NFC/RFID HARDWARE ===
# Scegliere UNA opzione in base al lettore hardware:

# OPZIONE A: Lettori USB/seriali generici
pyserial>=3.5

# OPZIONE B: Lettori NFC standard
pynfc>=0.3.3
# pyscard>=2.0.0

# OPZIONE C: Lettori RFID specifici  
# mfrc522>=0.0.7

# === PRODUZIONE ESSENZIALI ===
psutil>=5.8.0          # Monitoraggio sistema
schedule>=1.1.0        # Task scheduling automatici
watchdog>=2.1.0        # Monitoraggio file
requests>=2.25.0       # HTTP requests/API
cryptography>=3.4.0    # Sicurezza e encryption

# === EXPORT DATI ===
openpyxl>=3.0.0        # Export Excel
pandas>=1.3.0          # Analisi dati

# === NOTIFICHE ===
plyer>=2.1.0           # Notifiche sistema

# === PACKAGING DISTRIBUZIONE ===
pyinstaller>=4.5.0     # Crea eseguibili

# === WINDOWS SUPPORT ===
pywin32>=227           # Windows API (solo Windows)
wmi>=1.5.1             # Windows Management

# === DEVELOPMENT/TESTING ===
pytest>=6.2.0         # Testing framework
black>=21.0.0          # Code formatting

# Note: tkinter, datetime, threading, json, os, sys sono built-in Python
